+++ openpifpaf/transforms/annotations.py
@@ -1,0 +1,58 @@
+import copy
+import logging
+
+import numpy as np
+import torch
+
+from .preprocess import Preprocess
+
+LOG = logging.getLogger(__name__)
+
+
+class NormalizeAnnotations(Preprocess):
+    @staticmethod
+    def normalize_annotations(anns):
+        anns = copy.deepcopy(anns)
+
+        for ann in anns:
+            if 'keypoints' not in ann:
+                ann['keypoints'] = []
+
+            ann['keypoints'] = np.asarray(ann['keypoints'], dtype=np.float32).reshape(-1, 3)
+            ann['bbox'] = np.asarray(ann['bbox'], dtype=np.float32)
+            ann['bbox_original'] = np.copy(ann['bbox'])
+            if 'segmentation' in ann:
+                del ann['segmentation']
+
+        return anns
+
+    def __call__(self, image, anns, meta):
+        anns = self.normalize_annotations(anns)
+
+        if meta is None:
+            w, h = image.size
+            meta = {
+                'offset': np.array((0.0, 0.0)),
+                'scale': np.array((1.0, 1.0)),
+                'valid_area': np.array((0.0, 0.0, w, h)),
+                'hflip': False,
+                'width_height': np.array((w, h)),
+            }
+
+        return image, anns, meta
+
+
+class AnnotationJitter(Preprocess):
+    def __init__(self, epsilon=0.5):
+        self.epsilon = epsilon
+
+    def __call__(self, image, anns, meta):
+        meta = copy.deepcopy(meta)
+        anns = copy.deepcopy(anns)
+
+        for ann in anns:
+            keypoints_xy = ann['keypoints'][:, :2]
+            sym_rnd = (torch.rand(*keypoints_xy.shape).numpy() - 0.5) * 2.0
+            keypoints_xy += self.epsilon * sym_rnd
+
+        return image, anns, meta
