+++ openpifpaf/transforms/pad.py
@@ -1,0 +1,70 @@
+import copy
+import logging
+
+import torchvision
+
+from .preprocess import Preprocess
+
+LOG = logging.getLogger(__name__)
+
+
+class CenterPad(Preprocess):
+    def __init__(self, target_size):
+        self.log = logging.getLogger(__name__ + '.' + self.__class__.__name__)
+
+        if isinstance(target_size, int):
+            target_size = (target_size, target_size)
+        self.target_size = target_size
+
+    def __call__(self, image, anns, meta):
+        meta = copy.deepcopy(meta)
+        anns = copy.deepcopy(anns)
+
+        image, anns, ltrb = self.center_pad(image, anns)
+        meta['offset'] -= ltrb[:2]
+
+        self.log.debug('valid area before pad with %s: %s', ltrb, meta['valid_area'])
+        meta['valid_area'][:2] += ltrb[:2]
+        self.log.debug('valid area after pad: %s', meta['valid_area'])
+
+        for ann in anns:
+            ann['valid_area'] = meta['valid_area']
+
+        return image, anns, meta
+
+    def center_pad(self, image, anns):
+        w, h = image.size
+
+        left = int((self.target_size[0] - w) / 2.0)
+        top = int((self.target_size[1] - h) / 2.0)
+        if left < 0:
+            left = 0
+        if top < 0:
+            top = 0
+
+        right = self.target_size[0] - w - left
+        bottom = self.target_size[1] - h - top
+        if right < 0:
+            right = 0
+        if bottom < 0:
+            bottom = 0
+        ltrb = (left, top, right, bottom)
+
+        # pad image
+        image = torchvision.transforms.functional.pad(
+            image, ltrb, fill=(124, 116, 104))
+
+        # pad annotations
+        for ann in anns:
+            ann['keypoints'][:, 0] += ltrb[0]
+            ann['keypoints'][:, 1] += ltrb[1]
+            ann['bbox'][0] += ltrb[0]
+            ann['bbox'][1] += ltrb[1]
+
+        return image, anns, ltrb
+
+
+class SquarePad(Preprocess):
+    def __call__(self, image, anns, meta):
+        center_pad = CenterPad(max(image.size))
+        return center_pad(image, anns, meta)
