+++ openpifpaf/transforms/random.py
@@ -1,0 +1,15 @@
+import torch
+
+from .preprocess import Preprocess
+
+
+class RandomApply(Preprocess):
+    def __init__(self, transform, probability):
+        self.transform = transform
+        self.probability = probability
+
+    def __call__(self, image, anns, meta):
+        if float(torch.rand(1).item()) > self.probability:
+            return image, anns, meta
+        return self.transform(image, anns, meta)
+
