+++ setup.py
@@ -47,6 +47,7 @@ setup(
         'openpifpaf.decoder.generator',
         'openpifpaf.encoder',
         'openpifpaf.network',
+        'openpifpaf.transforms',
     ],
     license='GNU AGPLv3',
     description='PifPaf: Composite Fields for Human Pose Estimation',
